Move IApplicationXamlFileAccessor to its own file
Move `IApplicationXamlFileAccessor` to its own file, and put it in the Microsoft.VisualStudio.ProjectSystem.WPF namespace.
