+++ src/Microsoft.VisualStudio.ProjectSystem.Managed/ProjectSystem/Properties/InterceptedProjectProperties/ApplicationPropertyPage/WPFValueProvider.cs
@@ -1,5 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
+using Microsoft.VisualStudio.ProjectSystem.WPF;
+
 namespace Microsoft.VisualStudio.ProjectSystem.Properties;
 
 [ExportInterceptingPropertyValueProvider(
@@ -61,14 +63,4 @@ internal class WPFValueProvider : InterceptingPropertyValueProviderBase
             _ => throw new InvalidOperationException($"The {nameof(WPFValueProvider)} does not support the '{propertyName}' property.")
         };
     }
-}
-
-[ProjectSystemContract(ProjectSystemContractScope.UnconfiguredProject, ProjectSystemContractProvider.Extension, Cardinality = Composition.ImportCardinality.ExactlyOne)]
-internal interface IApplicationXamlFileAccessor
-{
-    Task<string?> GetStartupUriAsync();
-    Task SetStartupUriAsync(string startupUri);
-
-    Task<string?> GetShutdownModeAsync();
-    Task SetShutdownModeAsync(string shutdownMode);
 }
